battery/test_battery_basics.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Battery API</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Battery API **/

ok('battery' in navigator, "navigator.battery should exist");

var battery = navigator.battery;
is(battery.level, 0.5, "Emulator battery level should be 0.5")

is(battery.charging, true, "Emulator charging value should be true");
is(battery.dischargingTime, Infinity, "Emulator dischargingTime should be Inifinity");
is(battery.chargingTime, Infinity, "Emulator chargingTime should be Infinity");

</script>
</pre>
</body>
</html>
